Сделала плагин для систематизации фикстур и перенесла туда пару тестов для эластика
